Move appendFxMSTelemetryData to fetchData.ts
