+++ app/dashboard.tsx
@@ -4,11 +4,6 @@ import {
   fxmsMessageColumns,
 } from "./columns";
 import {
-  cleanLookerData,
-  mergeLookerData,
-  runLookQuery,
-} from "@/lib/looker.ts";
-import {
   _isAboutWelcomeTemplate,
 } from "../lib/messageUtils.ts";
 
@@ -50,26 +45,6 @@ export function compareDatesFn(a: NimbusRecipe, b: NimbusRecipe): number {
 
   // a must be equal to b
   return 0;
-}
-
-/**
- * Appends any FxMS telemetry message data from the query in Look
- * https://mozilla.cloud.looker.com/looks/2162 that does not already exist (ie.
- * no duplicate message ids) in existingMessageData and returns the result. The
- * message data is also cleaned up to match the message data objects from
- * ASRouter, remove any test messages, and update templates.
- */
-export async function appendFxMSTelemetryData(existingMessageData: any) {
-  // Get Looker message data (taken from the query in Look
-  // https://mozilla.cloud.looker.com/looks/2162)
-  const lookId = "2162";
-  let lookerData = await runLookQuery(lookId);
-
-  // Clean and merge Looker data with existing data
-  let jsonLookerData = cleanLookerData(lookerData);
-  let mergedData = mergeLookerData(existingMessageData, jsonLookerData);
-
-  return mergedData;
 }
 
 export async function getMsgRolloutCollection(
